adding benchmarks
