+++ benchmarks/run.py
@@ -1,0 +1,4 @@
+from benchmark_array_xd import benchmark_array_xd
+
+if __name__ == "__main__":
+    benchmark_array_xd()
